video info の contexts provider を component に切り出す
