+++ src/components/CommandExample.tsx
@@ -1,10 +1,10 @@
 import { css } from '@emotion/react';
 import { FormControlLabel, Switch, TextField, Tooltip } from '@mui/material';
 import { useContext, useState } from 'react';
-import { VideoInfoContext } from '../states/videoInfo';
 import { convertWinPathToWSL } from '../utils/convertWinPathToWSL';
 import { frameNumberToTimecode } from '../utils/frameNumberToTimecode';
 import { getOutPath } from '../utils/getOutPath';
+import { VideoInfoContext } from './VideoInfoContextsProvider';
 
 export const CommandExample = (props: {
   startFrameNumber: number;
