+++ src/components/FrameImage.tsx
@@ -2,8 +2,8 @@ import { css, SerializedStyles } from '@emotion/react';
 import { CircularProgress } from '@mui/material';
 import { useContext } from 'react';
 import { getFrameImageBase64InputSchema } from '../apiSchema';
-import { VideoInfoContext } from '../states/videoInfo';
 import { trpc } from '../trpc';
+import { VideoInfoContext } from './VideoInfoContextsProvider';
 
 export const FrameImage = (props: {
   frameNumber: number;
