+++ src/states/videoInfo.ts
@@ -1,5 +1,3 @@
-import { createContext } from 'react';
-
 export type VideoInfo =
   | {
       state: 'IDLE';
@@ -66,5 +64,3 @@ export const videoInfoReducer = (
       throw new Error('Unhandled action type');
   }
 };
-
-export const VideoInfoContext = createContext<VideoInfo>(videoInfoInitialState);
